requirements.txt
# TextCapture Python Dependencies
opencv-python
pytesseract
easyocr
pillow
numpy
pyperclip
PySide6
EasyOCR
pybind11